git-svn-id: svn://192.168.1.10/public/CoffeeMud@11275 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/core/CMParms.java
@@ -1590,7 +1590,7 @@ public class CMParms
 		}
 		final StringBuilder s=new StringBuilder(Integer.toString(V[0]));
 		for(int v=1;v<V.length;v++)
-			s.append(',').append((int)V[v]);
+			s.append(',').append(V[v]);
 		return s.toString();
 	}
